+++ canisters/rest-api/src/wallet/wallet.service.ts
@@ -4,36 +4,16 @@ import { Account } from 'azle/canisters/icrc_1/idl';
 
 import { ICP } from './tokens/icp';
 import { USDC_ARBITRUM } from './tokens/usdc_arbitrum';
+import { uuidToUint8Array } from './wallet.helpers';
 
 @Injectable()
 export class WalletService {
   constructor() {}
 
-  private stringToUint8Array(str: string): Uint8Array {
-    // Clean UUID by removing hyphens
-    const cleanStr = str.replace(/-/g, '');
-
-    const encoder = new TextEncoder();
-    const initialArray = encoder.encode(cleanStr);
-
-    // Ensure the input string is not longer than 32 bytes
-    if (initialArray.length > 32) {
-      throw new Error('Input string is too long - must be 32 bytes or less');
-    }
-
-    // Create a zero-filled array of exactly 32 bytes
-    const paddedArray = new Uint8Array(32).fill(0);
-
-    // Copy the encoded string bytes into the padded array
-    paddedArray.set(initialArray.slice(0, 32));
-
-    return paddedArray;
-  }
-
   private getAccount(userId: string): Account {
     return {
       owner: canisterSelf(),
-      subaccount: [this.stringToUint8Array(userId)],
+      subaccount: [uuidToUint8Array(userId)],
     };
   }
